requirements-docs.txt
app_version
django>=1.2
django-appconf
django-override-settings
mock
sphinx
sphinxcontrib-napoleon
